Move and rename files to .ts and update imports
