+++ packages/workbox-precaching/src/cleanupOutdatedCaches.ts
@@ -6,10 +6,10 @@
   https://opensource.org/licenses/MIT.
 */
 
-import {cacheNames} from 'workbox-core/_private/cacheNames.mjs';
-import {logger} from 'workbox-core/_private/logger.mjs';
-import {deleteOutdatedCaches} from './utils/deleteOutdatedCaches.mjs';
-import './_version.mjs';
+import {cacheNames} from 'workbox-core/_private/cacheNames.js';
+import {logger} from 'workbox-core/_private/logger.js';
+import {deleteOutdatedCaches} from './utils/deleteOutdatedCaches.js';
+import './_version.js';
 
 
 /**
