+++ packages/workbox-precaching/src/utils/cleanRedirect.ts
@@ -6,7 +6,7 @@
   https://opensource.org/licenses/MIT.
 */
 
-import '../_version.mjs';
+import '../_version.js';
 
 /**
  * @param {Response} response
