+++ packages/workbox-precaching/src/utils/deleteOutdatedCaches.ts
@@ -6,7 +6,7 @@
   https://opensource.org/licenses/MIT.
 */
 
-import '../_version.mjs';
+import '../_version.js';
 
 const SUBSTRING_TO_FIND = '-precache-';
