+++ packages/workbox-precaching/src/utils/getOrCreatePrecacheController.ts
@@ -6,8 +6,8 @@
   https://opensource.org/licenses/MIT.
 */
 
-import {PrecacheController} from '../PrecacheController.mjs';
-import '../_version.mjs';
+import {PrecacheController} from '../PrecacheController.js';
+import '../_version.js';
 
 
 let precacheController;
