+++ packages/workbox-precaching/src/utils/removeIgnoredSearchParams.ts
@@ -6,7 +6,7 @@
   https://opensource.org/licenses/MIT.
 */
 
-import '../_version.mjs';
+import '../_version.js';
 
 /**
  * Removes any URL search parameters that should be ignored.
